refactor(auth): move getTradeVolume method
