+++ lib/authenticated.js
@@ -1,6 +1,6 @@
 const PublicClient = require("./public.js");
 const SignRequest = require("./signer.js");
-const { API_LIMIT, _checkUndefined } = require("./utilities.js");
+const { _checkUndefined } = require("./utilities.js");
 
 class AuthenticatedClient extends PublicClient {
   /**
@@ -50,16 +50,6 @@ class AuthenticatedClient extends PublicClient {
       uri: request,
       headers: SignRequest({ key: this.key, secret: this.secret }, options)
     });
-  }
-
-  /**
-   * @example
-   * const volume = await authClient.getTradeVolume();
-   * @description Get trade volume for each symbol.
-   * @see {@link https://docs.gemini.com/rest-api/#get-trade-volume|get-trade-volume}
-   */
-  getTradeVolume() {
-    return this.post({ request: "/v1/tradevolume" });
   }
 
   /**
