+++ tests/authenticated.spec.js
@@ -37,72 +37,6 @@ suite("AuthenticatedClient", () => {
       .catch(error => assert.fail(error));
   });
 
-  test(".getTradeVolume()", done => {
-    const response = [
-      [
-        {
-          account_id: 5365,
-          symbol: "btcusd",
-          base_currency: "BTC",
-          notional_currency: "USD",
-          data_date: "2019-01-10",
-          total_volume_base: 8.06021756,
-          maker_buy_sell_ratio: 1,
-          buy_maker_base: 6.06021756,
-          buy_maker_notional: 23461.3515203844,
-          buy_maker_count: 34,
-          sell_maker_base: 0,
-          sell_maker_notional: 0,
-          sell_maker_count: 0,
-          buy_taker_base: 0,
-          buy_taker_notional: 0,
-          buy_taker_count: 0,
-          sell_taker_base: 2,
-          sell_taker_notional: 7935.66,
-          sell_taker_count: 2
-        },
-        {
-          account_id: 5365,
-          symbol: "ltcusd",
-          base_currency: "LTC",
-          notional_currency: "USD",
-          data_date: "2019-01-11",
-          total_volume_base: 3,
-          maker_buy_sell_ratio: 0,
-          buy_maker_base: 0,
-          buy_maker_notional: 0,
-          buy_maker_count: 0,
-          sell_maker_base: 0,
-          sell_maker_notional: 0,
-          sell_maker_count: 0,
-          buy_taker_base: 3,
-          buy_taker_notional: 98.22,
-          buy_taker_count: 3,
-          sell_taker_base: 0,
-          sell_taker_notional: 0,
-          sell_taker_count: 0
-        }
-      ]
-    ];
-    const request = "/v1/tradevolume";
-    const nonce = 1;
-    const payload = { request, nonce };
-    authClient.nonce = () => nonce;
-
-    nock(EXCHANGE_API_URL, { reqheaders: SignRequest(auth, payload) })
-      .post(request)
-      .times(1)
-      .reply(200, response);
-
-    authClient
-      .getTradeVolume()
-      .then(data => {
-        assert.deepStrictEqual(data, response);
-        done();
-      })
-      .catch(error => assert.fail(error));
-  });
-
   test(".newClearingOrder()", done => {
     const response = {
       result: "AwaitConfirm",
